test documentdb-session v1.0.0
