+++ package.json
@@ -40,7 +40,6 @@
     "applicationinsights": "^0.15.18",
     "cors": "^2.8.0",
     "documentdb": "^1.9.0",
-    "documentdb-session": "^1.0.0-alpha.4",
     "express": "^4.14.0",
     "express-handlebars": "^3.0.0",
     "express-session": "^1.14.0",
